Imbalance Algorithms

List of Algorithms Implemented

Imbalance.random_oversample

Imbalance.rose

Imbalance.smote

Imbalance.smoten

Imbalance.smotenc

Random Oversampler

Imbalance.random_oversample — Function

random_oversample(
    X, y; 
    ratios=nothing, rng=default_rng(), 
    try_preserve_type=true
)

Description

Naively oversample a dataset by randomly repeating existing observations with replacement.

Positional Arguments

X: A matrix or table of floats where each row is an observation from the dataset

y: An abstract vector of labels (e.g., strings) that correspond to the observations in X

Keyword Arguments

ratios=1.0: A parameter that controls the amount of oversampling to be done for each class

Can be a float and in this case each class will be oversampled to the size of the majority class times the float. By default, all classes are oversampled to the size of the majority class

Can be a dictionary mapping each class to the ratio of the needed number of observations for that class to the initial number of observations of the majority class

rng::Union{AbstractRNG, Integer}: Either an AbstractRNG object or an Integer seed to be used with Xoshiro

try_preserve_type::Bool=true: Defaults to true and means that the function will try to preserve the type of the input table (e.g., DataFrame). However, for some tables this may not succeed and in this case the table returned will be a column table (named-tuple of vectors).

Returns

Xover: A matrix or table that includes original data and the new observations due to oversampling. depending on whether the input X is a matrix or table respectively

yover: An abstract vector of labels that includes the original labels and the new instances of them due to oversampling

Example

using Imbalance
using StatsBase

# set probability of each class
probs = [0.5, 0.2, 0.3]                         
num_rows, num_continuous_feats = 100, 5
# generate a table and categorical vector accordingly
X, y = generate_imbalanced_data(num_rows, num_continuous_feats; 
                                probs, rng=42)                       
julia> StatsBase.countmap(y)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 33
1 => 19

# apply random oversampling
Xover, yover = random_oversample(X, y; ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
julia> StatsBase.countmap(yover)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 38
1 => 43

MLJ Model Interface

Simply pass the keyword arguments while initiating the RandomOversampler model and pass the positional arguments to the transform method.

using MLJ
RandomOversampler = @load RandomOversampler pkg=Imbalance

# Wrap the model in a machine
oversampler = RandomOversampler(ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
mach = machine(oversampler)

# Provide the data to transform (there is nothing to fit)
Xover, yover = transform(mach, X, y)

The MLJ interface is only supported for table inputs. Read more about the interface here.

TableTransforms Interface

This interface assumes that the input is one table Xy and that y is one of the columns. Hence, an integer y_ind must be specified to the constructor to specify which column y is followed by other keyword arguments. Only Xy is provided while applying the transform.

using Imbalance
using Imbalance.TableTransforms

# Generate imbalanced data
num_rows = 100
num_features = 5
y_ind = 3
Xy, _ = generate_imbalanced_data(num_rows, num_features; 
                                 probs=[0.5, 0.2, 0.3], insert_y=y_ind, rng=42)

# Initiate Random Oversampler model
oversampler = RandomOversampler(y_ind; ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
Xyover = Xy |> oversampler                    
Xyover, cache = TableTransforms.apply(oversampler, Xy)    # equivalently

The reapply(oversampler, Xy, cache) method from TableTransforms simply falls back to apply(oversample, Xy) and the revert(oversampler, Xy, cache) reverts the transform by removing the oversampled observations from the table.

ROSE

Imbalance.rose — Function

rose(
    X, y; 
    s=0.1, ratios=nothing, rng=default_rng(),
    try_perserve_type=true
)

Description

Oversamples a dataset using ROSE (Random Oversampling Examples) algorithm to correct for class imbalance as presented in [1]

Positional Arguments

X: A matrix or table of floats where each row is an observation from the dataset

y: An abstract vector of labels (e.g., strings) that correspond to the observations in X

Keyword Arguments

s::float: A parameter that proportionally controls the bandwidth of the Gaussian kernel

ratios=1.0: A parameter that controls the amount of oversampling to be done for each class

Can be a float and in this case each class will be oversampled to the size of the majority class times the float. By default, all classes are oversampled to the size of the majority class

Can be a dictionary mapping each class to the ratio of the needed number of observations for that class to the initial number of observations of the majority class

rng::Union{AbstractRNG, Integer}: Either an AbstractRNG object or an Integer seed to be used with Xoshiro

try_preserve_type::Bool=true: Defaults to true and means that the function will try to preserve the type of the input table (e.g., DataFrame). However, for some tables this may not succeed and in this case the table returned will be a column table (named-tuple of vectors).

Returns

Xover: A matrix or table that includes original data and the new observations due to oversampling. depending on whether the input X is a matrix or table respectively

yover: An abstract vector of labels that includes the original labels and the new instances of them due to oversampling

Example

using Imbalance
using StatsBase

# set probability of each class
probs = [0.5, 0.2, 0.3]                         
num_rows, num_continuous_feats = 100, 5
# generate a table and categorical vector accordingly
X, y = generate_imbalanced_data(num_rows, num_continuous_feats; 
                                probs, rng=42)                       
julia> StatsBase.countmap(y)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 33
1 => 19

# apply ROSE
Xover, yover = rose(X, y; s=0.3, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
julia> StatsBase.countmap(yover)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 38
1 => 43

MLJ Model Interface

Simply pass the keyword arguments while initiating the ROSE model and pass the positional arguments to the transform method.

using MLJ
ROSE = @load ROSE pkg=Imbalance

# Wrap the model in a machine
oversampler = ROSE(s=0.3, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
mach = machine(oversampler)

# Provide the data to transform (there is nothing to fit)
Xover, yover = transform(mach, X, y)

The MLJ interface is only supported for table inputs. Read more about the interface here.

TableTransforms Interface

This interface assumes that the input is one table Xy and that y is one of the columns. Hence, an integer y_ind must be specified to the constructor to specify which column y is followed by other keyword arguments. Only Xy is provided while applying the transform.

using Imbalance
using Imbalance.TableTransforms

# Generate imbalanced data
num_rows = 200
num_features = 5
y_ind = 3
Xy, _ = generate_imbalanced_data(num_rows, num_features; 
                                 probs=[0.5, 0.2, 0.3], insert_y=y_ind, rng=42)

# Initiate Random Oversampler model
oversampler = ROSE(y_ind; s=0.3, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
Xyover = Xy |> oversampler                              
Xyover, cache = TableTransforms.apply(oversampler, Xy)    # equivalently

The reapply(oversampler, Xy, cache) method from TableTransforms simply falls back to apply(oversample, Xy) and the revert(oversampler, Xy, cache) reverts the transform by removing the oversampled observations from the table.

References

[1] G Menardi, N. Torelli, “Training and assessing classification rules with imbalanced data,” Data Mining and Knowledge Discovery, 28(1), pp.92-122, 2014.

SMOTE

Imbalance.smote — Function

smote(
    X, y;
    k=5, ratios=nothing, rng=default_rng(),
    try_perserve_type=true
)

Description

Oversamples a dataset using SMOTE (Synthetic Minority Oversampling Techniques) algorithm to correct for class imbalance as presented in [1]

Positional Arguments

X: A matrix or table of floats where each row is an observation from the dataset

y: An abstract vector of labels (e.g., strings) that correspond to the observations in X

Keyword Arguments

k::Integer=5: Number of nearest neighbors to consider in the SMOTE algorithm. Should be within the range 0 < k < n where n is the number of observations in the smallest class. It will be automatically set to n-1 for any class where n ≤ k.

ratios=1.0: A parameter that controls the amount of oversampling to be done for each class

Can be a float and in this case each class will be oversampled to the size of the majority class times the float. By default, all classes are oversampled to the size of the majority class

Can be a dictionary mapping each class to the ratio of the needed number of observations for that class to the initial number of observations of the majority class

rng::Union{AbstractRNG, Integer}: Either an AbstractRNG object or an Integer seed to be used with Xoshiro

try_preserve_type::Bool=true: Defaults to true and means that the function will try to preserve the type of the input table (e.g., DataFrame). However, for some tables this may not succeed and in this case the table returned will be a column table (named-tuple of vectors).

Returns

Xover: A matrix or table that includes original data and the new observations due to oversampling. depending on whether the input X is a matrix or table respectively

yover: An abstract vector of labels that includes the original labels and the new instances of them due to oversampling

Example

using Imbalance
using StatsBase

# set probability of each class
probs = [0.5, 0.2, 0.3]                         
num_rows, num_continuous_feats = 100, 5
# generate a table and categorical vector accordingly
X, y = generate_imbalanced_data(num_rows, num_continuous_feats; 
                                probs, rng=42)                       
julia> StatsBase.countmap(y)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 33
1 => 19

# apply SMOTE
Xover, yover = smote(X, y; k = 5, ratios = Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng = 42)
julia> StatsBase.countmap(yover)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 38
1 => 43

MLJ Model Interface

Simply pass the keyword arguments while initiating the SMOTEN model and pass the positional arguments to the transform method.

using MLJ
SMOTE = @load SMOTE pkg=Imbalance

# Wrap the model in a machine
oversampler = SMOTE(k=5, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
mach = machine(oversampler)

# Provide the data to transform (there is nothing to fit)
Xover, yover = transform(mach, X, y)

The MLJ interface is only supported for table inputs. Read more about the interface here.

TableTransforms Interface

This interface assumes that the input is one table Xy and that y is one of the columns. Hence, an integer y_ind must be specified to the constructor to specify which column y is followed by other keyword arguments. Only Xy is provided while applying the transform.

using Imbalance
using Imbalance.TableTransforms

# Generate imbalanced data
num_rows = 200
num_features = 5
y_ind = 3
Xy, _ = generate_imbalanced_data(num_rows, num_features; 
                                 probs=[0.5, 0.2, 0.3], insert_y=y_ind, rng=42)

# Initiate Random Oversampler model
oversampler = SMOTE(y_ind; k=5, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
Xyover = Xy |> oversampler                              
Xyover, cache = TableTransforms.apply(oversampler, Xy)    # equivalently

The reapply(oversampler, Xy, cache) method from TableTransforms simply falls back to apply(oversample, Xy) and the revert(oversampler, Xy, cache) reverts the transform by removing the oversampled observations from the table.

References

[1] N. V. Chawla, K. W. Bowyer, L. O.Hall, W. P. Kegelmeyer, “SMOTE: synthetic minority over-sampling technique,” Journal of artificial intelligence research, 321-357, 2002.

SMOTE-N

Imbalance.smoten — Function

smoten(
    X, y;
    k, ratios=nothing, rng=default_rng(),
    try_perserve_type=true
)

Description

Oversamples a dataset using SMOTE-N (Synthetic Minority Oversampling Techniques-Nominal) algorithm to correct for class imbalance as presented in [1]. This is a variant of SMOTE to deal with datasets where all the features are nominal.

Positional Arguments

X: A matrix of integers or a table with elementscitypes that subtype Finite. That is, for table inputs each column should have either OrderedFactor or Multiclass as the element scitype.

y: An abstract vector of labels (e.g., strings) that correspond to the observations in X

Keyword Arguments

k::Integer=5: Number of nearest neighbors to consider in the SMOTE algorithm. Should be within the range 0 < k < n where n is the number of observations in the smallest class. It will be automatically set to n-1 for any class where n ≤ k.

ratios=1.0: A parameter that controls the amount of oversampling to be done for each class

Can be a float and in this case each class will be oversampled to the size of the majority class times the float. By default, all classes are oversampled to the size of the majority class

Can be a dictionary mapping each class to the ratio of the needed number of observations for that class to the initial number of observations of the majority class

rng::Union{AbstractRNG, Integer}: Either an AbstractRNG object or an Integer seed to be used with Xoshiro

try_preserve_type::Bool=true: Defaults to true and means that the function will try to preserve the type of the input table (e.g., DataFrame). However, for some tables this may not succeed and in this case the table returned will be a column table (named-tuple of vectors).

Returns

Xover: A matrix or table that includes original data and the new observations due to oversampling. depending on whether the input X is a matrix or table respectively

yover: An abstract vector of labels that includes the original labels and the new instances of them due to oversampling

Example

using Imbalance
using StatsBase

# set probability of each class
probs = [0.5, 0.2, 0.3]                         
num_rows = 100
num_continuous_feats = 0
# want two categorical features with three and two possible values respectively
cat_feats_num_vals = [3, 2]

# generate a table and categorical vector accordingly
X, y = generate_imbalanced_data(num_rows, num_continuous_feats; 
                                probs, cat_feats_num_vals, rng=42)                      
julia> StatsBase.countmap(y)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 33
1 => 19

julia> ScientificTypes.schema(X).scitypes
(Count, Count)

# coerce to a finite scitype (multiclass or ordered factor)
X = coerce(X, autotype(X, :few_to_finite))

# apply SMOTEN
Xover, yover = smoten(X, y; k=5, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
julia> StatsBase.countmap(yover)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 33
1 => 19

MLJ Model Interface

Simply pass the keyword arguments while initiating the SMOTEN model and pass the positional arguments to the transform method.

using MLJ
SMOTEN = @load SMOTEN pkg=Imbalance

# Wrap the model in a machine
oversampler = SMOTEN(k=5, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
mach = machine(oversampler)

# Provide the data to transform (there is nothing to fit)
Xover, yover = transform(mach, X, y)

The MLJ interface is only supported for table inputs. Read more about the interface here.

TableTransforms Interface

This interface assumes that the input is one table Xy and that y is one of the columns. Hence, an integer y_ind must be specified to the constructor to specify which column y is followed by other keyword arguments. Only Xy is provided while applying the transform.

using Imbalance
using ScientificTypes
using Imbalance.TableTransforms

# Generate imbalanced data
num_rows = 100
num_continuous_feats = 0
y_ind = 2
# generate a table and categorical vector accordingly
Xy, _ = generate_imbalanced_data(num_rows, num_continuous_feats; insert_y=y_ind,
                                probs= [0.5, 0.2, 0.3], cat_feats_num_vals=[3, 2],
                                 rng=42)  

# Table must have only finite scitypes                                
Xy = coerce(Xy, :Column1=>Multiclass, :Column2=>Multiclass, :Column3=>Multiclass)

# Initiate Random Oversampler model
oversampler = SMOTEN(y_ind; k=5, ratios=Dict(1=>1.0, 2=> 0.9, 3=>0.9), rng=42)
Xyover = Xy |> oversampler                               
Xyover, cache = TableTransforms.apply(oversampler, Xy)    # equivalently

The reapply(oversampler, Xy, cache) method from TableTransforms simply falls back to apply(oversample, Xy) and the revert(oversampler, Xy, cache) reverts the transform by removing the oversampled observations from the table.

References

[1] N. V. Chawla, K. W. Bowyer, L. O.Hall, W. P. Kegelmeyer, “SMOTE: synthetic minority over-sampling technique,” Journal of artificial intelligence research, 321-357, 2002.

SMOTE-NC

Imbalance.smotenc — Function

smotenc(
    X, y, split_ind;
    k=5, ratios=nothing, rng=default_rng(),
    try_perserve_type=true
)

Description

Oversamples a dataset using SMOTE-NC (Synthetic Minority Oversampling Techniques-Nominal Continuous) algorithm to correct for class imbalance as presented in [1]. This is a variant of SMOTE to deal with datasets with both nominal and continuous features.

SMOTE-NC Assumes Continuous Features Exist

SMOTE-NC will not work if the dataset is purely nominal. In that case, refer to SMOTE-N instead. Meanwhile, if the dataset is purely continuous then it's equivalent to the standard SMOTE`.

Positional Arguments

X: A matrix of floats or a table with element scitypes that subtype Union{Finite, Infinite}. Elements in nominal columns should subtype Finite (i.e., have scitype OrderedFactor or Multiclass) and elements in continuous columns should subtype Infinite (i.e., have scitype Count or Continuous).

y: An abstract vector of labels (e.g., strings) that correspond to the observations in X

cat_inds::AbstractVector{<:Int}: A vector of the indices of the nominal features. Supplied only if X is a matrix. Otherwise, they are inferred from the table's scitypes.

Keyword Arguments

k::Integer=5: Number of nearest neighbors to consider in the SMOTE algorithm. Should be within the range 0 < k < n where n is the number of observations in the smallest class. It will be automatically set to n-1 for any class where n ≤ k.

ratios=1.0: A parameter that controls the amount of oversampling to be done for each class

Can be a float and in this case each class will be oversampled to the size of the majority class times the float. By default, all classes are oversampled to the size of the majority class

Can be a dictionary mapping each class to the ratio of the needed number of observations for that class to the initial number of observations of the majority class

knn_tree: Decides the tree used in KNN computations. Either "Brute" or "Ball". BallTree can be much faster but may lead to innacurate results.

rng::Union{AbstractRNG, Integer}: Either an AbstractRNG object or an Integer seed to be used with Xoshiro

try_preserve_type::Bool=true: Defaults to true and means that the function will try to preserve the type of the input table (e.g., DataFrame). However, for some tables this may not succeed and in this case the table returned will be a column table (named-tuple of vectors).

Returns

Xover: A matrix or table that includes original data and the new observations due to oversampling. depending on whether the input X is a matrix or table respectively

yover: An abstract vector of labels that includes the original labels and the new instances of them due to oversampling

Example

using Imbalance
using StatsBase

# set probability of each class
probs = [0.5, 0.2, 0.3]                         
num_rows = 100
num_continuous_feats = 3
# want two categorical features with three and two possible values respectively
cat_feats_num_vals = [3, 2]

# generate a table and categorical vector accordingly
X, y = generate_imbalanced_data(num_rows, num_continuous_feats; 
                                probs, cat_feats_num_vals, rng=42)                      
julia> StatsBase.countmap(y)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 33
1 => 19

julia> ScientificTypes.schema(X).scitypes
(Continuous, Continuous, Continuous, Continuous, Continuous)
# coerce nominal columns to a finite scitype (multiclass or ordered factor)
X = coerce(X, :Column4=>Multiclass, :Column5=>Multiclass)

# apply SMOTE-NC
Xover, yover = smotenc(X, y; k = 5, ratios = Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng = 42)
julia>StatsBase.countmap(yover)
Dict{CategoricalArrays.CategoricalValue{Int64, UInt32}, Int64} with 3 entries:
0 => 48
2 => 33
1 => 19

MLJ Model Interface

Simply pass the keyword arguments while initiating the SMOTENC model and pass the positional arguments (excluding cat_inds) to the transform method.

using MLJ
SMOTEN = @load SMOTEN pkg=Imbalance

# Wrap the model in a machine
oversampler = SMOTENC(k=5, ratios=Dict(0=>1.0, 1=> 0.9, 2=>0.8), rng=42)
mach = machine(oversampler)

# Provide the data to transform (there is nothing to fit)
Xover, yover = transform(mach, X, y)

The MLJ interface is only supported for table inputs. Read more about the interface here.

TableTransforms Interface

This interface assumes that the input is one table Xy and that y is one of the columns. Hence, an integer y_ind must be specified to the constructor to specify which column y is followed by other keyword arguments. Only Xy is provided while applying the transform.

using Imbalance
using ScientificTypes
using Imbalance.TableTransforms

# Generate imbalanced data
num_rows = 100
num_continuous_feats = 3
y_ind = 2
# generate a table and categorical vector accordingly
Xy, _ = generate_imbalanced_data(num_rows, num_continuous_feats; insert_y=y_ind,
                                probs= [0.5, 0.2, 0.3], cat_feats_num_vals=[3, 2],
                                 rng=42)  

# Table must have only finite or continuous scitypes                                
Xy = coerce(Xy, :Column2=>Multiclass, :Column5=>Multiclass, :Column6=>Multiclass)

# Initiate Random Oversampler model
oversampler = SMOTENC(y_ind; k=5, ratios=Dict(1=>1.0, 2=> 0.9, 3=>0.9), rng=42)
Xyover = Xy |> oversampler                               
Xyover, cache = TableTransforms.apply(oversampler, Xy)    # equivalently

References

[1] N. V. Chawla, K. W. Bowyer, L. O.Hall, W. P. Kegelmeyer, “SMOTE: synthetic minority over-sampling technique,” Journal of artificial intelligence research, 321-357, 2002.